+++ Examples/Io/NuclearInteractions/include/ActsExamples/Io/NuclearInteractions/RootNuclearInteractionParametersWriter.hpp
@@ -9,7 +9,7 @@
 #pragma once
 
 #include "ActsExamples/Framework/WriterT.hpp"
-#include "ActsExamples/Io/Root/detail/NuclearInteractionParametrisation.hpp"
+#include "ActsExamples/Io/NuclearInteractions/detail/NuclearInteractionParametrisation.hpp"
 
 #include <mutex>
 #include <vector>
