+++ Examples/Run/HepMC3/CMakeLists.txt
@@ -14,7 +14,7 @@ target_link_libraries(
   ActsExampleHepMC3Histogram
   PRIVATE
     ActsCore ActsExamplesCommon
-    ActsExamplesFramework ActsExamplesIoHepMC3 ActsExamplesHepMC ActsExamplesIoRoot
+    ActsExamplesFramework ActsExamplesIoHepMC3 ActsExamplesHepMC ActsExamplesIoNuclearInteractions
     Boost::program_options)
     
 install(
